requirements.txt
# requirements.txt should contain only top-level requirements
# Update it manually: http://www.jfabre.com/how-not-to-to-pip-requirements/
# requirements.lock should be generated with pip freeze
Pillow==8.1.*
ImageHash==4.0
elasticsearch==6.3.1
